server/pyproject.toml
[tool.poetry]
name = "hammy-server"
version = "0.1.0"
description = ""
authors = ["Alex Ward <alxwrd@googlemail.com>"]
readme = "README.md"
packages = [{include = "hammy_server"}]

[tool.poetry.dependencies]
python = "^3.10"
python-socketio = "^5.10.0"


[tool.poetry.group.dev.dependencies]
uvicorn = "^0.24.0.post1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
